Make 'linkerd check' hint URLs dependent on version
The URLs have now the form `https://linkerd.io/checks/xx/#anchor` where
`xx` is extracted from the stable version string. If we're on an edge
version, or if it's a developer version such as `git-sha` or
`dev-sha-user`, then the `xx` will be `latest`.

These changes are coupled with linkerd/website#980