cli: show spinner for extension checks
Fixes #5637

Currently, When extension checks are running through `linkerd check`
we don't really any show status. This means that when the check
is run directly after a extension install, The CLI looks like it
got struck somewhere with no status as per the issue description.

This PR updates the code to show a spinner with the name of
extension of whose the check is being ran.

```bash

control-plane-version
---------------------
 control plane is up-to-date
 control plane and cli versions match
    control plane running stable-2.10.0 but cli running dev-e5ff84ce-tarun
    see https://linkerd.io/checks/#l5d-version-control for hints

Status check results are

Linkerd extensions checks
=========================

linkerd-jaeger
--------------
 linkerd-jaeger extension Namespace exists
 collector and jaeger service account exists
 collector config map exists
 jaeger extension pods are injected
 jaeger extension pods are running

Status check results are

/ Running viz extension check
```

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>