Remove nginx container from mc gateway pod
Fixes #5444

Replaced the nginx container with a pause container, and removed all the
config boilerplate associated to nginx.

The probe now targets the proxy's readiness endpoint, port 4191 path
`/ready`.

## Upgrade note

When multicluster gets updated in both source and target this won't
cause downtime because this change only affects the probing. The `Links`
will have to be re-generated of course. Til that is done, `linkerd mc
gateways` will show the target cluster as not alive, but that shouldn't
affect exiting connections.