cli: restrict repair cmd with the same server version
Fixes #5917

Currently, When a repair cmd is ran with a CLI that is different
from the server version. This means it is possible for the cmd
to fail as the `Values` struct changes between versions.

This PR updates the cmd to restrict it to the version that the
server is on, but as the repair cmd is avialable only on `2.9.4`
we fall-back on to that version in the suggestion when the server
version is `2.9` or 2.8`.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>