Update bin/web script to work with metrics-api (#5944)
This change updates the bin/web script to work with linkerd viz's
`metrics-api`. Previously, the script was still pointing to the
linkerd `controller` pod which no longer has the public-api to get
metrics.

This change also fixes the script's port forward function so that we can
point to both the `linkerd-controller` and the `metrics-api`. This was
causing the linkerd web development environment to not start up when
running the script.

Signed-off-by: Dennis Adjei-Baah <dennis@buoyant.io>