jaeger: fix prometheus scrape configuration
Fixes #5976

Currently, Jaeger and Collector components in jaeger extension
do not actually support metrics scraping because relevant
ports are not exposed and prometheus annotations are not set
correctly.

This PR fixes those values to be the correct ones.

By default, prometheus in `linkerd-viz` does not actually
scrape jaeger metrics, and additional configuration
has to be applied to do the same.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>