cli: make inject add the right annotation with `--ingress` (#5983)
* cli: make inject add the right annotation with `--ingress`

Currently, When `--ingress` flag is used with inject. The
CLI is not setting `linkerd.io/inject: ingress`. Instead,
we get a `linkerd.io/inject: enabled`.

This is because we are setting the annotation to `ingress` in
`getOverrideAnnotations` which is being overridden to `enabled`
later in the inject pipeline at `transform`. This PR updates
the `transform` to only add the `enabled` annotation if its
not ingress mode injection.

After this change:

```bash
 ./bin/go-run cli inject https://gist.githubusercontent.com/Pothulapati/941eef18a862aee9adb805aeddaef844/raw/005f4facb3b5805c48b0baf3f480f1fc6fd67f27/shell.yaml --ingress --proxy-cpu-limit 50m
github.com/linkerd/linkerd2/cli/cmd
github.com/linkerd/linkerd2/cli
apiVersion: v1
kind: Pod
metadata:
  annotations:
    config.linkerd.io/proxy-cpu-limit: 50m
    linkerd.io/inject: ingress
  name: shell-demo
spec:
  containers:
  - image: nginx
    name: nginx
    volumeMounts:
    - mountPath: /usr/share/nginx/html
      name: shared-data
  dnsPolicy: Default
  volumes:
  - emptyDir: {}
    name: shared-data
---

pod "shell-demo" injected
```

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>