Make HTTP/2 Keepalive ping intervals configurable
Some HTTP/2 services are very finicky and particular about
the timing of pings. In particular, Go's implementation of gRPC
requires that clients and servers exactly match up the timing of their
keepalive pings. If pings are off for more than MAX_PING_STRIKES pings,
the connection will be closed via ENHANCE YOUR CALM. The linkerd-proxy
pod spec currently hardcodes the keepalive timing at 10s, which is fine
for fully meshed services, but presents a problem when connecting
meshed services to unmeshed services. Namely, that connections with
differing keepalive intervals will be closed after some period of time
(specific to the underlying keepalive algorithms) and clients dropped
with an UNAVAILABLE error (h2 translates the GOAWAY into `excessive
load balancing behavior`, which then translates into UNAVAILABLE
within linkerd-proxy).

The solution to this is to make the keepalive interval configurable by
those who need to connect unmeshed services to meshed ones. The
environment variable is already exposed on the sidecare container spec,
so it is just a matter of surfacing it up.

The change can be validated through the normal testing infrastructure,
or by creating a minimal go server with a keepalive setting different
from that of an ingress proxy and toggling the keepalive settings on
both to either be different or match.

Fixes #5988

Signed-off-by: Jack Andersen <jandersen@plaid.com>