Removed Destination's `Get` API from the public-api
This is the first step towards removing the `linkerd-controller` pod. It deals with removing the Destination `Get` http and gRPC endpoint it exposes, that only the `linkerd diagnostics endpoints` is consuming.

Removed all references to Destination in the public-api, including all the gRPC-to-http-to-gRPC forwardings:
- Removed the `Get` method from the public-api gRPC server that forwarded the connection from the controller pod to the destination pod. Clients should now connect directly to the destination service via gRPC.
- Likewise, removed the destination boilerplate in the public-api http server (and its `main.go`) that served the `Get` destination endpoint and forwarded it into the gRPC server.
- Finally, removed the destination boilerplate from the public-api's `client.go` that created a client connecting to the http API.

In order to connect to the `Get` API, the `endpoints` command now calls a `GetDestinationClient` method in `api.go` that runs a series of healthchecks.
The healthchecks have a new entry "can initialize the destination client" that builds the destination client, by calling a new function `NewExternalClient` defined in `controller/api/destination/client.go` that port-forwards the destination service before creating a gRPC client pointing to it.

This change will be followed-up by the replacement of the `Version` API, and another separate change for completely removing the `linkerd-controller` once all its clients have been dealt with.