trafficsplit: Support v1alpha2 version
This PR updates the TrafficSplit CRD to add the `v1alpha2` version
which changes the `weight` field from `resource.Quantity` into
integer.

Our Initial approach to fixing this, was to switch all the types in
the code repo to be `v1alpha2`. This was complex because switching
from resource.Quantity to integer means that a conversion webhook
would be needed to support existing users and not break stuff in
clusters.

Instead of changing all golang TrafficSplit types to `v1alpha2`, This
PR tries to do it differently i.e Keep using `storage` version and
all the types in destination as `v1alpha1` and make k8s convert `v1alpha2`
resources into `v1alpha1`. This works because `v1alpha1`'s `weight` field
is `resource.Quantity` which also covers normal integers
(which is in v1alpha2). By keeping all the internal types to still use
`v1lapha1`, we can essentially support both types without having to write
conversion webhooks. The obv trade-off here is that we are essentially
not moving away from `v1alpha1` in the code repository.

Because with TrafficSplit, The proportions matter more than the exact digits
we are not loosing any type of precision when converting from a `integer`
into `resource.Quantity`. So, With this change, The following TrafficSplit
resource should work.

```yaml
apiVersion: split.smi-spec.io/v1alpha2
kind: TrafficSplit
metadata:
  name: reviews-rollout
spec:
  service: reviews
  backends:
  - service: reviews-v2
    weight: 50
  - service: reviews-v3
    weight: 50
```

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>