Change web base image from debian to scratch
It only required removing `RUN mkdir -p app` from the Dockerfile, which
is not really needed.

This diminishes the image size and avoids having frequent vulnerability
reports by image scanners in clusters appearing in debian. We were
upgrading the base image whenever we could, but it's an issue for older
releases.

Addresses comment in #5198