consolidate Indexers into Server, to prevent confusion across watchers
This commit moves the indexer creations into Server.go, so that there
is a single place where all indexers are created instead of each watcher
doing its own. This works because the same `k8s.API` is used across watchers
and the server.

With this change, The golang watcher tests are expected to fail as the informers
are outside of the watcher logic that they test.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>