Add test for opaque traffic between meshed and unmeshed
Opaque ports should still work on an unmeshed service. Currently, there
is no integration test to validate that traffic between a meshed app and
an unmeshed (but opaque) service works as expected.

This change adds a new integration test that confirms opaque ports
behaves correctly in the given scenario. Correct behavior in this case
means that there will only be TCP metrics available on the inbound side
(and not on the outbound, unmeshed side). Furthermore, no HTTP metrics
should be recorded by either side (as consequence of opaqueness).

Signed-off-by: Matei David <matei@buoyant.io>