add `TestTrafficSplitCliWithSP` integration test
This commit adds a new integration test to the set of TrafficSplit
integration tests that we already have to check for TrafficSplitting
through `serviceProfiles.dstOverrides`

This integration test follows the same pattern of other traffic
split integration tests i.e first applying a trafficsplit with
weights `500m` & `0`, followed by `500m` and `500m` and checking
if the traffic is distributed.

This integration does not check in the same manner as `linkerd
viz stat ts` does not yet work with service profiles with
`dstOverrides`. Instead, we perform `linkerd viz stat` on the
client and check if there are expected server objects to be
there.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>