Introduce support for StatefulSets across multicluster
The purpose of this PR is to prototype (and demonstrate) stateful sets
across multicluster. Currently, it isn't possible to communicate with a
specific pod in a statefulset across clusters without manually creating
clusterIP services for each pod backing the statefulset in the target
cluster.

By allowing the service mirror to automatically create those clusterIP
services on behalf of the user, we can achieve a scalable and automated
way of communicating directly to a pod across clusters. We do so by
relying on "service redirection" whereby a headless service is
replicated in the source cluster without a clusterIP. Along with the
headless service, the service mirror will create a clusterIP service for
each pod in the set (with its Endpoints object pointing to the Gateway).
The service mirror creates the root service's Endpoints object with each
clusterIP service paired with its respective host in the target cluster.

To exemplify, assume a statefulset "foo" in a target cluster 'west' with
2 pods (foo-0, foo-1). In the src cluster 'east', we create a headless
  'foo-west' and 2 services (foo-0-west, foo-1-west) whose Endpoints
  point to the Gateway IP.  Foo-west's Endpoints will contain an
  AddressSet with two hosts:
  - hostname: foo-0 ip: <clusterIP of foo-0-west>
  - hostname: foo-1 ip: <clusterIP of foo-1-west>

To achieve this "redirection", the service mirror needs to distinguish
between a normal exported service and a headless exported service.
Additionally, the destination service should resolve a clusterIP service
marked as "headless" to one Endpoint (weighted addr set) in GetProfile.

Signed-off-by: Matei David <matei@buoyant.io>