Fix check repeatedly printing error on retry (#6082)
#6038 refactored `healthcheck_output.go` to accomodate the printing
check results in `short` output. In that refactor the spinner display
was code section was changed so that that code could be shared in the
different `prettyPrintResults` functions.

Previously, the function would exit early prior to printing to avoid
showing the the check error output if the check was undergoing a retry.
After the refactor, that code was inadvertently removed causing check
to always print the error message and then continue retrying.

This caused output like this:
```
kubernetes-version
------------------
 is running the minimum Kubernetes API version
 is running the minimum kubectl version

linkerd-existence
-----------------
 'linkerd-config' config map exists
 heartbeat ServiceAccount exist
 control plane replica sets are ready
 no unschedulable pods
 control plane pods are ready
    pod/linkerd-identity-7cbd8bcbdf-hlmdw container linkerd-proxy is not
ready
    see https://linkerd.io/checks/#l5d-api-control-ready for hints
 control plane pods are ready
    No running pods for "linkerd-proxy-injector"
    see https://linkerd.io/checks/#l5d-api-control-ready for hints
 control plane pods are ready
    No running pods for "linkerd-proxy-injector"
    see https://linkerd.io/checks/#l5d-api-control-ready for hints
 control plane pods are ready
    No running pods for "linkerd-proxy-injector"
    see https://linkerd.io/checks/#l5d-api-control-ready for hints
 control plane pods are ready
    No running pods for "linkerd-proxy-injector"
    see https://linkerd.io/checks/#l5d-api-control-ready for hints
 control plane pods are ready
    No running pods for "linkerd-proxy-injector"
    see https://linkerd.io/checks/#l5d-api-control-ready for hints
```

This change fixes the issue by forcing the print function to exit early
to avoid the repeated output message. We still want the message to be
printed once, only after check has given up retrying. This change should
handle that case.

Fixes #6080

Signed-off-by: Dennis Adjei-Baah <dennis@buoyant.io>