Versioned `linkerd check` hint URLs (#6102)
Supersedes #5885

We've recently refactored the linkerd docs in the website to host separate sections for each major 2.x version. The [Troubleshooting page](https://linkerd.io/2.10/tasks/troubleshooting/) section is also versioned, but currently the hint URLs output by `linkerd check` commands point to `https://linkerd.io/checks` which redirects to `https://linkerd.io/2.10/tasks/troubleshooting`.

This PR changes the base of those hint URL so that instead it points to `https://linkerd.io/{version}/checks` where `{version}` is the CLI's major version. If the CLI is not in a stable version (edge or local dev version) then `{version}` will be `2` which the site will redirect to the latest stable version.

This change is coupled with linkerd/website#5885 that takes care of the website-side of things.