Introduce Endpoints watcher and diff
Previously, scaling endpoints wouldn't have any effect on StatefulSet
pods that have been mirrored. Through this change we can now reconcile
the desired state in the target cluster with the actual state in the
source cluster.

Also, we now use Endpoints instead of pods to get hostnames for
StatefulSet pods that are exported as services. Added more debug
statements and simplified some of the logic.

Signed-off-by: Matei David <matei@buoyant.io>