update `smi-adaptor` to ignore specific SP's
Currently, There is no way to prevent `smi-adaptor` from making
changes to a specific `SP`. This is useful for users and other
controllers which want to take control.

This commit adds a new `smi.linkerd.io/skip` annotation which when
annotated to a `SP` prevents the `smi-adaptor` from mutating it.

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>