Add nodePorts option to multicluster helm chart (#6059)
* Add nodePorts option and group gateway options

For LoadBalancer and NodePort type services it should be possible to set
a specific node port, this wasn't possible yet. Also, the gateway options
were not grouped nicely.

This commit adds nodePorts to the multicluster helm chart. Also, it groups
all gateway options under the gateway group.

Note; this is now a breaking change

Signed-off-by: Peter Smit <peter.smit@inscripta.io>