Change how we check for root annotation length
In the proxy injector, when we want to inherit an opaque ports
annotation value from a namespace for a service, we can generate a patch
with the root annotations map and the value, and a patch only with the
value. At the moment, the patch generation logic looks at the
resourceConfig pod's annotation when deciding whether the annotations
map should be created. This causes an issue with services that have
existing annotations -- since the logic checks the pod's annotation, the
generated patch will always include a root annotations map which
overwrites all of the service's existing annotations.

This change introduces a fix by looking at the resourceConfig's workload
annotations instead of a pod specifically. Through this change, we
should be able to determine more accurately whether the patch should
also create an annotations map.

Signed-off-by: Matei David <matei@buoyant.io>