Move Endpoints creation logic to createOrUpdateEndpoints
Previously, when creating a headless mirror we would also create the
Endpoints object. This approach does not seem to work very well; if a
service is exported when it is created, its backing pods will not show
up in the corresponding Endpoints logic.

To fix this, Endpoints creation for a headless service has been moved to
a different function and it is triggered by updates picked up by the
informer. Other changes in this commit include refreshed comments to
describe the situation better.

Signed-off-by: Matei David <matei@buoyant.io>