remove the need for `updated-for` annotation
Currently, We cleanup only the SP that is updated or created by the
`smi-adaptor`. We can instead cleanup all the time for a consistent
experience.

The thinking here is that, whenever the `smi-adaptor` is present,
the users loose control over `sp.dstOverrides`.

For this to work, The workqueue has been updated to store
also `ts.spec.service` so that we can find the relevant
sp and delete

Signed-off-by: Tarun Pothulapati <tarunpothulapati@outlook.com>